# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Teya Ledger API
  description: "Event-sourced money ledger: customers, accounts, deposits, withdrawals,\
    \ transaction history. All write endpoints require an Idempotency-Key header."
  contact:
    name: Teya
  license:
    name: MIT
  version: 1.0.0
servers:
- url: http://localhost:18080
  description: Generated server url
...
paths:
...
  /account/{accountId}/transaction:
    get:
      tags:
      - Transaction
      summary: List money-movement history for an account
      description: "Cursor-paginated. nextCursor is null when no further records exist.\
        \ Lifecycle events (account opened/closed, overdraft changed) are not surfaced\
        \ here."
      operationId: history
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: string
      - name: after
        in: query
        required: false
        schema:
          type: integer
          format: int64
          default: 0
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 50
      responses:
        "200":
          description: Page of transactions
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransactionPageResponse"
        "400":
          description: "Invalid limit (must be in [1, 200])"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /account/{accountId}/balance:
    get:
      tags:
      - Transaction
      summary: Point-in-time balance for an account
      description: Returns the balance after applying every money event with occurredAt
        <= at. The cutoff is inclusive. Unknown accounts return 0 (consistent with
        empty history). Lifecycle events are ignored — they don't move money.
      operationId: balanceAt
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: string
      - name: at
        in: query
        required: true
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: Balance at the requested instant
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BalanceResponse"
        "400":
          description: Missing or malformed 'at' parameter
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties:
            type: object
        requestId:
          type: string
...
    Item:
      type: object
      properties:
        seq:
          type: integer
          format: int64
        type:
          type: string
        amountMinorUnits:
          type: integer
          format: int64
        currency:
          type: string
        occurredAt:
          type: string
          format: date-time
    TransactionPageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        nextCursor:
          type: integer
          format: int64
    BalanceResponse:
      type: object
      properties:
        balanceMinorUnits:
          type: integer
          format: int64
        asOf:
          type: string
          format: date-time